Little cleanup between Positions/Coordinates utility classes
