+++ src/java/org/tensorics/core/tensor/Position.java
@@ -30,124 +30,128 @@ import java.util.HashSet;
 import java.util.Set;
 
 import com.google.common.base.Preconditions;
-import com.google.common.collect.ClassToInstanceMap;
 import com.google.common.collect.ImmutableMultiset;
 import com.google.common.collect.Interner;
 import com.google.common.collect.Interners;
 
 /**
- * Defines the position of a value within a tensor in the N-dimensional coordinate space.
+ * Defines the position of a value within a tensor in the N-dimensional
+ * coordinate space.
  * 
  * @author agorzaws
  */
 public final class Position implements Serializable {
 
-    private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
 
-    private static final Interner<Position> INTERNER = Interners.newWeakInterner();
+	private static final Interner<Position> INTERNER = Interners.newWeakInterner();
 
-    /*
-     * NOTE: This has to be after the cachedPositions initialization, because the 'of' method uses it!
-     */
-    private static final Position EMPTY_POSITION = Position.createFrom(Collections.emptySet());
+	/*
+	 * NOTE: This has to be after the cachedPositions initialization, because
+	 * the 'of' method uses it!
+	 */
+	private static final Position EMPTY_POSITION = Position.createFrom(Collections.emptySet());
 
-    private Set<?> coordinates;
+	private Set<?> coordinates;
 
-    private Position(Set<?> coordinates) {
-        this.coordinates = new HashSet<>(coordinates);
-        for (Object oneCoordinate : coordinates) {
-            if (oneCoordinate instanceof Position) {
-                throw new IllegalArgumentException("A position is contained in the collection of coordinates."
-                        + "This is most-probably a programming mistake and therefore not allowed.");
-            }
-        }
-    }
+	private Position(Set<?> coordinates) {
+		this.coordinates = new HashSet<>(coordinates);
+		for (Object oneCoordinate : coordinates) {
+			if (oneCoordinate instanceof Position) {
+				throw new IllegalArgumentException("A position is contained in the collection of coordinates."
+						+ "This is most-probably a programming mistake and therefore not allowed.");
+			}
+		}
+	}
 
-    @SuppressWarnings("PMD.ShortMethodName")
-    public static Position of(Set<?> coordinates) {
-        return createFrom(requireValidCoordinates(coordinates));
-    }
+	@SuppressWarnings("PMD.ShortMethodName")
+	public static Position of(Set<?> coordinates) {
+		return createFrom(requireValidCoordinates(coordinates));
+	}
 
-    private static Position createFrom(Set<?> coordinates) {
-        return INTERNER.intern(new Position(coordinates));
-    }
+	private static Position createFrom(Set<?> coordinates) {
+		return INTERNER.intern(new Position(coordinates));
+	}
 
-    public static Position empty() {
-        return EMPTY_POSITION;
-    }
+	public static Position empty() {
+		return EMPTY_POSITION;
+	}
 
-    @SafeVarargs
-    @SuppressWarnings("PMD.ShortMethodName")
-    public static Position of(Object... coordinates) {
-        return createFrom(requireValidCoordinates(ImmutableMultiset.copyOf(coordinates)));
-    }
+	@SafeVarargs
+	@SuppressWarnings("PMD.ShortMethodName")
+	public static Position of(Object... coordinates) {
+		return createFrom(requireValidCoordinates(ImmutableMultiset.copyOf(coordinates)));
+	}
 
-    public Set<Object> getCoordinates() {
-        return new HashSet<>(coordinates);
-    }
+	public Set<Object> getCoordinates() {
+		return new HashSet<>(coordinates);
+	}
 
-    public <CS> CS coordinateFor(Class<CS> dimension) {
-        return Positions.firstCoordinateOfTyp(coordinates, dimension);
-    }
+	public <CS> CS coordinateFor(Class<CS> dimension) {
+		return Coordinates.firstCoordinateOfTyp(coordinates, dimension);
+	}
 
-    public Set<?> coordinates() {
-        return new HashSet<>(coordinates);
-    }
+	public Set<?> coordinates() {
+		return new HashSet<>(coordinates);
+	}
 
-    /**
-     * Retrieves the dimensions of this position (i.e. the type of the containing coordinates)
-     * 
-     * @return the types of the coordinates
-     */
-    public Set<Class<?>> dimensionSet() {
-        return Positions.getDimensionsFrom(coordinates);
-    }
+	/**
+	 * Retrieves the dimensions of this position (i.e. the type of the
+	 * containing coordinates)
+	 * 
+	 * @return the types of the coordinates
+	 */
+	public Set<Class<?>> dimensionSet() {
+		return Coordinates.getDimensionsFrom(coordinates);
+	}
 
-    /**
-     * Checks if the position is consistent with the given dimensions. Conformity means that the position contains
-     * exactly one coordinate for each dimension in the given set of dimensions (classes of coordinates).
-     * 
-     * @param dimensions the dimensions for which conformity has to be checked.
-     * @return {@code true} if the position is conform, {@code false} if not.
-     */
-    public boolean isConsistentWith(Set<? extends Class<?>> dimensions) {
-        Preconditions.checkArgument(dimensions != null, "Argument '" + "dimensions" + "' must not be null!");
-        return Positions.areDimensionsConsistentWithCoordinates(dimensions, this);
-    }
+	/**
+	 * Checks if the position is consistent with the given dimensions.
+	 * Conformity means that the position contains exactly one coordinate for
+	 * each dimension in the given set of dimensions (classes of coordinates).
+	 * 
+	 * @param dimensions
+	 *            the dimensions for which conformity has to be checked.
+	 * @return {@code true} if the position is conform, {@code false} if not.
+	 */
+	public boolean isConsistentWith(Set<? extends Class<?>> dimensions) {
+		Preconditions.checkArgument(dimensions != null, "Argument '" + "dimensions" + "' must not be null!");
+		return Positions.areDimensionsConsistentWithCoordinates(dimensions, this);
+	}
 
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((coordinates == null) ? 0 : coordinates.hashCode());
-        return result;
-    }
+	@Override
+	public int hashCode() {
+		final int prime = 31;
+		int result = 1;
+		result = prime * result + ((coordinates == null) ? 0 : coordinates.hashCode());
+		return result;
+	}
 
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        Position other = (Position) obj;
-        if (coordinates == null) {
-            if (other.coordinates != null) {
-                return false;
-            }
-        } else if (!coordinates.equals(other.coordinates)) {
-            return false;
-        }
-        return true;
-    }
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj) {
+			return true;
+		}
+		if (obj == null) {
+			return false;
+		}
+		if (getClass() != obj.getClass()) {
+			return false;
+		}
+		Position other = (Position) obj;
+		if (coordinates == null) {
+			if (other.coordinates != null) {
+				return false;
+			}
+		} else if (!coordinates.equals(other.coordinates)) {
+			return false;
+		}
+		return true;
+	}
 
-    @Override
-    public String toString() {
-        return "Position [coordinates=" + coordinates + "]";
-    }
+	@Override
+	public String toString() {
+		return "Position [coordinates=" + coordinates + "]";
+	}
 
 }
